Created reusable `ListViewComposerFactory` in `EssentialApp`
