+++ EssentialApp/EssentialApp/FeedUIComposer.swift
@@ -7,7 +7,11 @@ public final class FeedUIComposer {
     public static func composeWith(loader: @escaping () -> AnyPublisher<[FeedImage], Error>, imageLoader: @escaping (URL) -> FeedImageLoader.Publisher) -> ListViewController {
         let adapter = LoadResourcePresentationAdapter<[FeedImage], FeedViewAdapter>(loader: { loader().dispatchOnMainQueue() })
         
-        let viewController = ListViewController.makeWith(title: FeedPresenter.title, onRefresh: adapter.loadResource)
+        let viewController = ListViewController.makeWith(
+            title: FeedPresenter.title,
+            onRefresh: adapter.loadResource,
+            storyboardName: "Feed"
+        )
         let view = FeedViewAdapter(imageLoader: { imageLoader($0).dispatchOnMainQueue() })
         view.controller = viewController
 
@@ -22,14 +26,3 @@ public final class FeedUIComposer {
     }
 }
 
-private extension ListViewController {
-    static func makeWith(title: String, onRefresh: @escaping () -> Void) -> ListViewController {
-        let bundle = Bundle(for: ListViewController.self)
-        let storyboard = UIStoryboard(name: "Feed", bundle: bundle)
-        let controller = storyboard.instantiateInitialViewController() as! ListViewController
-        controller.title = title
-        controller.onRefresh = onRefresh
-
-        return controller
-    }
-}
