+++ EssentialApp/EssentialApp/ImageCommentsUIComposer.swift
@@ -7,7 +7,11 @@ public class ImageCommentsUIComposer {
     public static func composeWith(loader: @escaping () -> AnyPublisher<[ImageComment], Error>) -> ListViewController {
         let adapter = LoadResourcePresentationAdapter<[ImageComment], ImageCommentsViewAdapter>(loader: { loader().dispatchOnMainQueue() })
 
-        let viewController = ListViewController.makeWith(title: ImageCommentsPresenter.title, onRefresh: adapter.loadResource)
+        let viewController = ListViewController.makeWith(
+            title: ImageCommentsPresenter.title,
+            onRefresh: adapter.loadResource,
+            storyboardName: "ImageComments"
+        )
         let view = ImageCommentsViewAdapter()
         view.controller = viewController
 
@@ -19,17 +23,5 @@ public class ImageCommentsUIComposer {
         )
 
         return viewController
-    }
-}
-
-private extension ListViewController {
-    static func makeWith(title: String, onRefresh: @escaping () -> Void) -> ListViewController {
-        let bundle = Bundle(for: ListViewController.self)
-        let storyboard = UIStoryboard(name: "ImageComments", bundle: bundle)
-        let controller = storyboard.instantiateInitialViewController() as! ListViewController
-        controller.title = title
-        controller.onRefresh = onRefresh
-
-        return controller
     }
 }
